requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
bidict==0.22.1
click==8.1.7
fastapi==0.105.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
inflection==0.5.1
looseversion==1.3.0
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.8.1
python-multipart==0.0.6
python-socketio==5.10.0
rethinkdb==2.4.10.post1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.25.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4